+++ example/spotify_dart_example.dart
@@ -4,19 +4,31 @@
 import 'dart:io';
 import 'dart:convert';
 import '../lib/spotify_io.dart';
-import '../lib/spotify_io.json.g.dart';
 
 main() async {
-  var keyJson = await new File('.apikeys').readAsString();
+  var keyJson = await new File('example/.apikeys').readAsString();
   var keyMap = JSON.decode(keyJson);
 
   var credentials = new SpotifyApiCredentials(keyMap['id'], keyMap['secret']);
   var spotify = new SpotifyApi(credentials);
 
+  print("Artists:");
   var artists = await spotify.artists.list(['0OdUWJ0sBjDrqHygGUXeCF']);
-
   artists.forEach((x) => print(x.name));
 
-  var album = await spotify.albums.get('6PEYvIrLYHJ8BvpE5uUChz');
+  print("Album:");
+  var album = await spotify.albums.get('2Hog1V8mdTWKhCYqI5paph');
   print(album.name);
+
+  print("Album Tracks:");
+  var tracks = await spotify.albums.getTracks(album.id).all();
+  tracks.forEach((track) {
+    print(track.name);
+  });
+
+  print("Featured Playlists:");
+  var featuredPlaylists = await spotify.playlists.featured.all();
+  featuredPlaylists.forEach((playlist) {
+    print(playlist.name);
+  });
 }