+++ lib/src/spotify_browser.dart
@@ -10,8 +10,13 @@ class SpotifyApi extends SpotifyApiBase {
   Future<String> _getImpl(String url, Map<String, String> headers) async {
     var client = new http.BrowserClient();
     var response = await client.get(url, headers: headers);
-
-    return UTF8.decode(response.bodyBytes);
+    var responseBody = UTF8.decode(response.bodyBytes);
+    if (response.statusCode >= 400) {
+      var json = JSON.decode(responseBody);
+      throw new SpotifyException.fromSpotify(
+          SpotifyErrorMapper.parse(json['error']));
+    }
+    return responseBody;
   }
 
   @override
@@ -19,6 +24,26 @@ class SpotifyApi extends SpotifyApiBase {
       String url, Map<String, String> headers, dynamic body) async {
     var client = new http.BrowserClient();
     var response = await client.post(url, headers: headers, body: body);
-    return UTF8.decode(response.bodyBytes);
+    var responseBody = UTF8.decode(response.bodyBytes);
+    if (response.statusCode >= 400) {
+      var json = JSON.decode(responseBody);
+      throw new SpotifyException.fromSpotify(
+          SpotifyErrorMapper.parse(json['error']));
+    }
+    return responseBody;
+  }
+
+  @override
+  Future<String> _putImpl(
+      String url, Map<String, String> headers, dynamic body) async {
+    var client = new http.BrowserClient();
+    var response = await client.put(url, headers: headers, body: body);
+    var responseBody = UTF8.decode(response.bodyBytes);
+    if (response.statusCode >= 400) {
+      var json = JSON.decode(responseBody);
+      throw new SpotifyException.fromSpotify(
+          SpotifyErrorMapper.parse(json['error']));
+    }
+    return responseBody;
   }
 }