+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.0.7
+
+- add user's saved tracks endpoints
+
+## 0.0.6
+
+- add pagination support
+
+## 0.0.5
+
+- add user and playlist endpoints
+
 ## 0.0.4
 
 - fix typo in artists endpoint